// src/lib/webApi/generated/models/clientResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * GoPadel.Admin
 * OpenAPI spec version: v1
 */

export interface ClientResponse {
  id?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  surname?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  level?: number | null;
  /** @nullable */
  rating?: number | null;
}