// src/lib/webApi/generated/models/clientResponsePaginatedResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * GoPadel.Admin
 * OpenAPI spec version: v1
 */
import type { ClientResponse } from './clientResponse';

export interface ClientResponsePaginatedResponse {
  /** @nullable */
  items?: ClientResponse[] | null;
  page?: number;
  count?: number;
  totalCount?: number;
  totalPages?: number;
  hasNextPage?: boolean;
  hasPreviousPage?: boolean;
}